+++ src/configs/import.ts
@@ -53,7 +53,7 @@ export interface ImportEslintConfigOptions extends UnConfigOptions<'import'> {
 
 export const importUnConfig: UnConfigFn<'import'> = async (context) => {
   const [eslintPluginImportX, {createTypeScriptImportResolver}] = await Promise.all([
-    pluginsLoaders.import() as unknown as Promise<typeof import('eslint-plugin-import-x')>,
+    pluginsLoaders.import(context) as unknown as Promise<typeof import('eslint-plugin-import-x')>,
     interopDefault(import('eslint-import-resolver-typescript')),
   ]);